SolucionDemostrable/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==47.0.0
et_xmlfile==2.0.0
fastapi==0.136.1
h11==0.16.0
idna==3.13
lxml==6.1.0
openpyxl==3.1.5
packaging==26.2
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pypdfium2==5.7.1
pytesseract==0.3.13
python-multipart==0.0.26
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
